feat: rework track module to interact with database

+++ src/track/track.service.ts
@@ -1,16 +1,11 @@
-import {
-  Inject,
-  Injectable,
-  NotFoundException,
-  UnprocessableEntityException,
-  forwardRef,
-} from '@nestjs/common';
-import { v4 as uuidV4 } from 'uuid';
+import { Inject, Injectable, NotFoundException, forwardRef } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
 
 import { CreateTrackDto, UpdateTrackDto } from './dto';
 import { FavsService } from 'src/favs/favs.service';
-import { ArtistService } from 'src/artist/artist.service';
-import { AlbumService } from 'src/album/album.service';
+
+import { Track } from './entities/track.entity';
 
 @Injectable()
 export class TrackService {
@@ -18,98 +13,44 @@ export class TrackService {
     @Inject(forwardRef(() => FavsService))
     private readonly favsService: FavsService,
 
-    @Inject(forwardRef(() => ArtistService))
-    private readonly artistService: ArtistService,
-
-    @Inject(forwardRef(() => AlbumService))
-    private readonly albumService: AlbumService,
+    @InjectRepository(Track)
+    private trackRepository: Repository<Track>,
   ) {}
 
-  private tracks: { [id: string]: Track } = {};
-
-  private findTrack(id: string): Track {
-    const track = this.tracks[id];
+  private async findTrack(id: string) {
+    const track = await this.trackRepository.findOneBy({ id });
     if (!track) throw new NotFoundException('Track not found');
 
     return track;
   }
 
-  private async checkAlbumAndArtistExistence(albumId: string | null, artistId: string | null) {
-    if (albumId) {
-      try {
-        await this.albumService.findOne(albumId);
-      } catch {
-        throw new UnprocessableEntityException(`Album with id ${albumId} not found`);
-      }
-    }
+  create(createTrackDto: CreateTrackDto) {
+    const createdTrack = this.trackRepository.create(createTrackDto);
 
-    if (artistId) {
-      try {
-        await this.artistService.findOne(artistId);
-      } catch {
-        throw new UnprocessableEntityException(`Artist with id ${artistId} not found`);
-      }
-    }
+    return this.trackRepository.save(createdTrack);
   }
 
-  async create(createTrackDto: CreateTrackDto) {
-    const id = uuidV4();
-    const { albumId, artistId, duration, name } = createTrackDto;
-
-    await this.checkAlbumAndArtistExistence(albumId, artistId);
-
-    const newTrack: Track = { albumId, artistId, duration, id, name };
-
-    this.tracks[id] = newTrack;
-
-    return newTrack;
+  async findAll() {
+    return (await this.trackRepository.find()).map((track) => track.toResponse());
   }
 
-  findAll() {
-    return Object.values(this.tracks);
-  }
-
-  findOne(id: string) {
-    return this.findTrack(id);
+  async findOne(id: string) {
+    return (await this.findTrack(id)).toResponse();
   }
 
   async update(id: string, updateTrackDto: UpdateTrackDto) {
-    this.findTrack(id);
+    const track = await this.findTrack(id);
 
-    const { albumId, artistId, duration, name } = updateTrackDto;
+    Object.assign(track, updateTrackDto);
 
-    await this.checkAlbumAndArtistExistence(albumId, artistId);
-
-    const updatedTrack = { id, albumId, artistId, duration, name };
-
-    this.tracks[id] = updatedTrack;
-
-    return updatedTrack;
+    return this.trackRepository.save(track);
   }
 
-  remove(id: string) {
-    this.findTrack(id);
-    this.tracks = Object.fromEntries(Object.entries(this.tracks).filter(([key]) => key !== id));
+  async remove(id: string) {
+    const deleteResult = await this.trackRepository.delete(id);
+
+    if (!deleteResult.affected) throw new NotFoundException('Track not found');
+
     this.favsService.removeTrack(id);
-  }
-
-  removeArtistFromTracks(artistId: string) {
-    this.tracks = Object.fromEntries(
-      Object.entries(this.tracks).map(([trackId, track]) => {
-        if (track.artistId === artistId) return [trackId, { ...track, artistId: null }];
-
-        return [trackId, track];
-      }),
-    );
-  }
-
-  removeAlbumFromTracks(albumId: string) {
-    this.tracks = Object.fromEntries(
-      Object.entries(this.tracks).map(([trackId, track]) => {
-        if (track.albumId === albumId) return [trackId, { ...track, albumId: null }];
-
-        return [trackId, track];
-      }),
-    );
   }
 }
